--- Scripts/Runtime/Jobs/AccessControl/AbstractAccessController.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 7a044aa53b2746bcb9516809d0e9f20e
-timeCreated: 1650651401
