+++ Scripts/Runtime/Jobs/AccessControl/AccessType.cs
@@ -1,7 +1,7 @@
 namespace Anvil.Unity.DOTS.Jobs
 {
     /// <summary>
-    /// Represents the type of access to a collection for use with a <see cref="AbstractAccessController"/>
+    /// Represents the type of access to a collection for use with a <see cref="AccessController"/>
     /// </summary>
     public enum AccessType
     {
